// generated/src/app/api-models/agent-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AgentMaster {
  seqAgentId: number;
  agentId: string;
  agentType: string;
  lastName: string;
  firstName: string;
  middleInitial: string;
  shortName: string;
  dateOfBirth: Date;
  levelNumber: number;
  seqParentAgentId: number;
  backupWitholding: string;
  commissionCalcThruDate: Date;
  commissionPaidThruDate: Date;
  commissionActivity: string;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}